[HOT-FIX] fixed some implicit sign conversions
Former-commit-id: 031e3370d0bbe2e030530b0fb1a4c582747a72ea
Former-commit-id: a9200a6a5a8b77099bf1247d6f9dedbb835cc301

+++ src/simulation/utilities/gauss_markov.h
@@ -43,7 +43,7 @@ class GaussMarkov
     
 public:
     GaussMarkov();
-    GaussMarkov(uint64_t size);
+    GaussMarkov(uint32_t size);
     ~GaussMarkov();
     void computeNextState();
     /*!@brief Method does just what it says, seeds the random number generator
